requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
